🎨 Moved business logic to services
