+++ middleware/app/api/routes/health.py
@@ -1,20 +1,11 @@
+import api.services.health as health_service
 import utils.logger as logger
-from database.db import DynamoDBManager
 from fastapi import APIRouter
-from storage.cloudflare_r2 import CloudflareR2Manager
 
 router = APIRouter()
 
 # Logger instance
 log = logger.get_logger()
-
-# DynamoDB
-table_name = "byteshare-upload-metadata"
-dynamodb = DynamoDBManager(table_name)
-
-# Storage
-BUCKET_NAME = "byteshare-blob"
-storage = CloudflareR2Manager(BUCKET_NAME)
 
 
 @router.get("/")
@@ -32,8 +23,7 @@ def health_check():
     FUNCTION_NAME = "health_check()"
     log.info("Entering {}".format(FUNCTION_NAME))
 
-    dynamodb.health_check()
-    storage.health_check()
+    response = health_service.health_check()
 
     log.info("Exiting {}".format(FUNCTION_NAME))
-    return {"status": "ok", "details": "Service is running"}
+    return response
